+++ ts/components/staking/wizard/number_input.tsx
