+++ ts/components/staking/wizard/transaction_item.tsx
